// frontend/src/types/api.ts
export interface ApiResponse<T> {
  success: boolean
  message?: string
  data: T
}

export interface PaginationMeta {
  page: number
  pageSize: number
  total: number
  totalPages: number
}

export interface PaginatedResponse<T> {
  success: boolean
  message?: string
  data: T[]
  meta: PaginationMeta
}